refactor(ipc): extract vocabulary + replacement commands into peer module (#431)
First slice of the ipc/commands/mod.rs file split. Lifts the 8
CRUD commands for the Replacements + Vocabulary subsystems out of
the 3,300-LOC mega-module into a new
src/ipc/commands/dictionary.rs peer, matching the existing pattern
of meeting.rs / models.rs / macos.rs / export.rs.

Both subsystems share the IpcError::Replacements variant by
design (the frontend renders them under one Dictionary mental
model), so they live in the same peer rather than separating
into two modules.

Registration in lib.rs uses the full path
(ipc::commands::dictionary::vocabulary_list etc.) per the
pub-use-loses-the-macro-symbol trap documented in CLAUDE.md and
learnings.md 2026-04-25.

mod.rs: 3,315 → ~3,200 LOC. The remaining sub-task split (history,
settings, dictation, ptt, updater) ships in follow-up PRs.

Refs #431.
