+++ src-tauri/src/lib.rs
@@ -492,14 +492,14 @@ pub fn run() {
             ipc::commands::history_count,
             ipc::commands::history_clear,
             ipc::commands::get_dictation_stats,
-            ipc::commands::replacements_list,
-            ipc::commands::replacement_create,
-            ipc::commands::replacement_update,
-            ipc::commands::replacement_delete,
-            ipc::commands::vocabulary_list,
-            ipc::commands::vocabulary_create,
-            ipc::commands::vocabulary_update,
-            ipc::commands::vocabulary_delete,
+            ipc::commands::dictionary::replacements_list,
+            ipc::commands::dictionary::replacement_create,
+            ipc::commands::dictionary::replacement_update,
+            ipc::commands::dictionary::replacement_delete,
+            ipc::commands::dictionary::vocabulary_list,
+            ipc::commands::dictionary::vocabulary_create,
+            ipc::commands::dictionary::vocabulary_update,
+            ipc::commands::dictionary::vocabulary_delete,
             ipc::commands::models::model_list,
             ipc::commands::models::model_select,
             ipc::commands::models::model_download,
